+++ p2panda-rs/src/test_utils/fixtures/params.rs
@@ -19,15 +19,6 @@ use crate::test_utils::constants::{DEFAULT_HASH, DEFAULT_PRIVATE_KEY, DEFAULT_SC
 use crate::test_utils::fixtures::defaults;
 use crate::test_utils::utils;
 
-/// Fixture struct which contains versioned p2panda data for testing.
-#[derive(Debug)]
-pub struct Fixture {
-    pub entry: Entry,
-    pub entry_signed_encoded: EntrySigned,
-    pub key_pair: KeyPair,
-    pub operation_encoded: OperationEncoded,
-}
-
 /// Fixture which injects the default private key string into a test method.
 #[fixture]
 pub fn private_key() -> String {
@@ -201,35 +192,4 @@ pub fn encoded_create_string(create_operation: Operation) -> String {
         .unwrap()
         .as_str()
         .to_owned()
-}
-
-/// Fixture which injects p2panda testing data from p2panda version `0.3.0`.
-#[fixture]
-pub fn v0_3_0_fixture() -> Fixture {
-    let operation_fields = utils::operation_fields(vec![
-        ("name", OperationValue::Text("chess".to_string())),
-        (
-            "description",
-            OperationValue::Text("for playing chess".to_string()),
-        ),
-    ]);
-    let operation = create_operation(Hash::new(DEFAULT_SCHEMA_HASH).unwrap(), operation_fields);
-    let key_pair = utils::keypair_from_private(
-        "4c21b14046f284f87f1ea4be4b973664221ad483079a68ed35a6812553b41176".into(),
-    );
-
-    // Comment out to regenerate fixture:
-    // let entry_signed_encoded = sign_and_encode(
-    //     &entry(operation.clone(), seq_num(1), None, None),
-    //     &key_pair,
-    // ).unwrap();
-    // println!("{:?}", entry_signed_encoded.as_str());
-    // println!("{?}", OperationEncoded::try_from(&operation)).unwrap();
-
-    Fixture {
-        entry_signed_encoded: EntrySigned::new("009cdb3a8c0c4b308173d4c3c43a67a6d013444af99acb8be6c52423746d9aa2c10101b6002065c34e1997b82fd08fc886bf6c2803cfaf93e3ad4da9128a661eb79e30f97bee25e525e72c99394ec91c33195f6b43c78274bd3096938260d5e18f237b57211d0a8e9eee49f594c1ddfb609ee0f9d0f502bf8701c3b2e5b0c34c61ec3e614a02").unwrap(),
-        operation_encoded: OperationEncoded::new("a466616374696f6e6663726561746566736368656d61784430303230633635353637616533376566656132393365333461396337643133663866326266323364626463336235633762396162343632393331313163343866633738626776657273696f6e01666669656c6473a26b6465736372697074696f6ea26474797065637374726576616c756571666f7220706c6179696e67206368657373646e616d65a26474797065637374726576616c7565656368657373").unwrap(),
-        key_pair,
-        entry: entry(operation, seq_num(1), None, None)
-    }
 }